requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

numpy==1.26.4
opencv-python==4.9.0.80
python-dotenv==1.0.1
